전체적인 Log, 주석 지우기 및 DataSource 분리
