+++ app/src/main/java/com/example/accountbook/presentation/MainActivity.kt
@@ -2,15 +2,11 @@ package com.example.accountbook.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.viewModels
 import androidx.core.view.doOnLayout
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.commit
 import com.example.accountbook.R
-import com.example.accountbook.data.db.AccountBookDbHelper
-import com.example.accountbook.data.model.Categories
-import com.example.accountbook.data.model.Payments
 import com.example.accountbook.databinding.ActivityMainBinding
 import com.example.accountbook.presentation.calendar.CalendarFragment
 import com.example.accountbook.presentation.history.HistoryDetailFragment
@@ -49,9 +45,8 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun initBottomNavigation() {
-        binding.mainBottomNavView.doOnLayout { 
+        binding.mainBottomNavView.doOnLayout {
             settingViewModel.bottomNavigationHeight = it.height
-            Log.e(TAG, "initBottomNavigation: ${it.height}", )
         }
         if (supportFragmentManager.fragments.isEmpty()) {
             setFragment(mainViewModel.curSelectedMenuItemId.value!!)
@@ -59,9 +54,9 @@ class MainActivity : AppCompatActivity() {
 
         binding.mainBottomNavView.setOnItemSelectedListener {
 
-            with(mainViewModel){
+            with(mainViewModel) {
                 isDeleteMode.value?.let { deleteMode ->
-                    if (deleteMode){
+                    if (deleteMode) {
                         resetDeleteModeProperties()
                     }
                 }
@@ -78,17 +73,16 @@ class MainActivity : AppCompatActivity() {
 
     private fun setFragment(menuItemId: Int) {
         var fragment = supportFragmentManager.findFragmentByTag(menuItemId.toString())
-        Log.e(TAG, "setFragment: backstackCount: ${supportFragmentManager.backStackEntryCount}", )
         supportFragmentManager.popBackStack()
-        if (fragment == null){
-            when(menuItemId){
+        if (fragment == null) {
+            when (menuItemId) {
                 R.id.bottom_nav_item_history -> fragment = HistoryFragment()
                 R.id.bottom_nav_item_calendar -> fragment = CalendarFragment()
                 R.id.bottom_nav_item_statistics -> fragment = StatisticsFragment()
                 R.id.bottom_nav_item_setting -> fragment = SettingFragment()
             }
-        }else{
-            when(menuItemId){
+        } else {
+            when (menuItemId) {
                 R.id.bottom_nav_item_history -> {
                     fragment as HistoryFragment
                 }
@@ -110,16 +104,12 @@ class MainActivity : AppCompatActivity() {
     }
 
     override fun onBackPressed() {
-        with(supportFragmentManager){
-            if (backStackEntryCount!=2){
+        with(supportFragmentManager) {
+            if (backStackEntryCount != 2) {
                 historyDetailViewModel.resetMemberProperties()
-            }else{
-                //popBackStack()
             }
-            Log.e(TAG, "onBackPressed: $backStackEntryCount", )
-            Log.e(TAG, "onBackPressed: ${this.fragments}", )
-            if(this.fragments.first() !is HistoryFragment){
-                if (fragments.size==1 && fragments.last() !is HistoryDetailFragment && fragments.last() !is SettingDetailFragment){
+            if (this.fragments.first() !is HistoryFragment) {
+                if (fragments.size == 1 && fragments.last() !is HistoryDetailFragment && fragments.last() !is SettingDetailFragment) {
                     binding.mainBottomNavView.selectedItemId = R.id.bottom_nav_item_history
                     setFragment(R.id.bottom_nav_item_history)
                     return
