+++ app/src/main/java/com/example/accountbook/presentation/calendar/CalendarFragment.kt
@@ -2,11 +2,7 @@ package com.example.accountbook.presentation.calendar
 
 import android.os.Build
 import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.GridLayoutManager
